There is now a separate function to create the access token payload for each type (one for users and another for clients).
The system now identifies whether a user or a client is accessing the query, granting or denying access accordingly. This is scalable to more different types using an enum.

The previous auth repository has been renamed to refresh_token. A new auth repository has been created that retrieves both users and clients, used for get_current_principal.

The unit_of_work is now injected into the services.

Added a /me route in the clients module.